Fix viewmodel flip on zoom
The fabs() was incorrectly applied only to the viewmodel's FOV instead
of the result of that subtracted by the FOV offset.

Further it doesn't seem to make sense to use the absolute value of that
subtraction; Yes, it does prevent flipping, but it will zoom out again
as the FOV decreases.
Instead just limit the result of the subtraction to non-negative
numbers.